+++ src/Universum/Lifted.hs
@@ -1,17 +1,22 @@
 {-# LANGUAGE Safe #-}
 
+{-
+Copyright: (c) 2016 Stephen Diehl
+           (c) 20016-2018 Serokell
+           (c) 2018 Kowainik
+License: MIT
+-}
+
 -- | Lifted versions of base functions.
 
 module Universum.Lifted
        ( module Universum.Lifted.Concurrent
-       , module Universum.Lifted.Env
+       , module Universum.Lifted.Exit
        , module Universum.Lifted.File
        , module Universum.Lifted.IORef
-       , module Universum.Lifted.ST
        ) where
 
 import Universum.Lifted.Concurrent
-import Universum.Lifted.Env
+import Universum.Lifted.Exit
 import Universum.Lifted.File
 import Universum.Lifted.IORef
-import Universum.Lifted.ST